% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mardia.R
\name{mardia}
\alias{mardia}
\title{Univariate and Multivariate skewness and kurtosis checker}
\usage{
mardia(x, na.rm = TRUE)
}
\arguments{
\item{x}{A data matrix}

\item{na.rm}{An indication of the missing data, the default value is True}
}
\value{
Data information: sample size and number of variables. The marginal and multivariate test (Mardia's Test) of skewness and kurtosis.
}
\description{
Univariate and Multivariate skewness and kurtosis checker
}
